# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/internal/processors/platform/client/ClientRequestHandler.java
#	modules/core/src/main/java/org/apache/ignite/internal/processors/platform/client/ClientResponse.java

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/platform/client/cache/ClientCacheScanQueryNextPageResponse.java
@@ -33,7 +33,7 @@ class ClientCacheScanQueryNextPageResponse extends ClientResponse {
      * @param requestId Request id.
      * @param cursor Cursor.
      */
-    ClientCacheScanQueryNextPageResponse(int requestId, ClientCacheScanQueryCursor cursor) {
+    ClientCacheScanQueryNextPageResponse(long requestId, ClientCacheScanQueryCursor cursor) {
         super(requestId);
 
         assert cursor != null;
